+++ packages/orq-rc/RELEASES.md
@@ -10359,3 +10359,13 @@ Based on:
 - [typescript v4.11.0-rc.57] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.57] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.57 - packages/orq-rc
+
+## 2026-06-30 06:08:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.786.1 (2.913.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.58] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.58] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.58 - packages/orq-rc